Changed Path and Player Clicks
Made it so when the path is being shown it stays green for the whole path.

When the play clicks on a square not on the path the game ends.

The player must click the squares in the correct order.

The player cannot click other squares while the path is being shown.

Hitting an enemy will kill you.

If the player dies the same path will be shown. The player will do the same path until they get it complete.